api: wip fix things

+++ api/src/modules/barbershops/use-cases/delete-blocked-date.use-case.ts
@@ -1,10 +1,13 @@
 import { BarbershopRepository } from "../gateways/barbershop.repository";
-import { BarbershopNotFoundError } from "../errors/barbershop-errors";
+import { BlockedDateNotFoundError } from "../errors/barbershop-errors";
 
 export class DeleteBlockedDateUseCase {
   constructor(private barbershopRepository: BarbershopRepository) {}
 
   async execute(blockedDateId: string) {
+    const blockedDate = await this.barbershopRepository.findBlockedDateById(blockedDateId);
+    if (!blockedDate) throw new BlockedDateNotFoundError();
+
     await this.barbershopRepository.deleteBlockedDate(blockedDateId);
   }
 }
